os-watch/pyproject.toml
[tool.poetry]
name = "os-watch"
version = "0.1.0"
description = ""
authors = ["BexTuychiev <bex@ibexprogramming.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.43.2"
pydantic = "^2.10.6"
firecrawl-py = "^1.14.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
